Fix double submit button bug for password creation
**Why**:

* Problem: When I then add a new valid password and click submit, the
  "weak password" error message goes away, but the new password is not
  submitted and the submit button must be clicked again.
* This is more of a design issue than anything. When the validation
  errors are removed, it shifts the button up so if you happened to
  click anywhere on the bottom half of the button the click isn't
  actually registered on the submit button... top half of button will
  submit successfully
* Note: this PR removing client-side validation on "change" -- we hope
  to add it back in but was too much of a hassle to do at the moment.
* Still have client-side validation on submit
