+++ app/assets/javascripts/application.js
@@ -1,4 +1,3 @@
-
 import 'app/pw-toggle';
 import 'app/form-validation';
 import 'app/form-field-format';
